sched: fair: Do not boost cpu/task utilization if boost = 2
For the same reasons as a6d616a6ab860156d98d37563a2bb2ece0a116be, but still allowing a user
configuration if a user does want the extra performance the corresponding
power costs.

Choose a boost value of 2 for this as it's not typically set as much as
something like 1, 5, 10, etc. would be used.

Signed-off-by: idkwhoiam322 <idkwhoiam322@raphielgang.org>
Signed-off-by: atndko <z1281552865@gmail.com>
Signed-off-by: alanndz <alanndz7@gmail.com>