phy: msm: usb: put phy in normal mode during pulsing for CDP
Commit 6b06fff3c102 ("phy: msm: usb: Drive pulse on D+ data line
in case of CDP") added support to apply pulse on D+ data line in case
of CDP to avoid transition of host port from CDP to SDP. As part of
that, it is required to put phy in normal mode before driving pulse,
but not putting PHY in normal mode which results in pulse not seen
by host. Hence set PHY mode to normal driving mode by writing correct
value to opmode field of PHY register to fix issue.

Change-Id: I1376b987235a87a7f35a3934025ab3599ffca5b2
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: alanndz <alanndz7@gmail.com>